services/worker/requirements.txt
nats-py==2.7.2
redis==5.0.1
prometheus_client==0.20.0
